Implement AbstractMutableHashConsEGraph and refactor HashConsEGraphBuilder to extend it
